Split mgdata from objsub
Mario Party 8 suggests this split.
