+++ config/GP5E01/symbols.txt
@@ -1091,7 +1091,7 @@ THPViewFunc = .text:0x80061778; // type:function size:0xA4 scope:local
 THPViewSprFunc = .text:0x8006181C; // type:function size:0x17C scope:local
 THPDecodeFunc = .text:0x80061998; // type:function size:0x4C scope:local
 HuTHPMatHook = .text:0x800619E4; // type:function size:0x654
-omOvlMgNoGet = .text:0x80062038; // type:function size:0x5C
+MgNoGet = .text:0x80062038; // type:function size:0x5C
 MgSubModeGet = .text:0x80062094; // type:function size:0x8
 MgDecaScoreCalc = .text:0x8006209C; // type:function size:0x458
 omGameSysInit = .text:0x800624F4; // type:function size:0x114
